Fix the vertex offset in indices
The offsets in the meshlets are still broken though.

+++ app/src/ModelBase.cpp
@@ -71,7 +71,7 @@ void ModelBundleBase::SetModels(float modelScale, const MeshBundleImpl& meshBund
 				std::make_shared<ModelBase>(modelScale)
 			);
 
-			model->SetMeshIndex(static_cast<std::uint32_t>(currentNodeData.meshIndex));
+			model->SetMeshIndex(currentNodeData.meshIndex);
 		}
 	}
 }
